% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{args_env}
\alias{args_env}
\title{create an environment for args}
\usage{
args_env(..., parent = parent.frame())
}
\arguments{
\item{...}{objects}

\item{parent}{parent environment}
}
\description{
create an environment for args
}
