% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{callwith}
\alias{callwith}
\title{Evaluate a function with a modified default values}
\usage{
callwith(fun, args, dots = list(), keep.null = FALSE, envir = parent.frame())
}
\arguments{
\item{fun}{either a function or a non-empty character string naming the function to be called}

\item{args}{a list of values to modify the default arguments of the function}

\item{dots}{the user-specific input (usually from ...)}

\item{keep.null}{\code{TRUE} to keep \code{NULL} values after argument modifications}

\item{envir}{the environment to evaluate the function call}
}
\description{
Evaluate a function with a modified default values
}
