% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comparers}
\alias{comparers}
\alias{equal}
\alias{exact}
\alias{include}
\alias{like}
\alias{pattern}
\title{Comparer functions}
\usage{
exact(x, y = get(".data", envir = parent.frame()), ...)

equal(x, y = get(".data", envir = parent.frame()))

include(x, y = get(".data", envir = parent.frame()))

pattern(x, y = get(".data", envir = parent.frame()), ...)

like(x, dist = 1L, y = get(".data", envir = parent.frame()), ...)
}
\arguments{
\item{x}{target}

\item{y}{source}

\item{dist}{maximum distance}

\item{...}{additional parameters:
for \code{exact}, passed to \code{identical};
for \code{pattern}, passed to \code{grepl};
for \code{like}, passed to \code{stringdist::stringdist}.}
}
\value{
\code{logical}
}
\description{
Comparer functions
}
\details{
\code{exact}: test if two objects are exactly identical.

\code{equal}: test if two atomic vectors of the same mode and length
   are equal.

\code{include}: test if the values of an atomic vector are included
   by the other with the same mode, respectively.

\code{pattern}: test if the values match a particular pattern of
   regular expression.

\code{like}: test if the distance between two atomic character vectors
   is no greater than a given value.
}

