% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contains}
\alias{contains}
\title{Test if a vector contains certain values @param table the values to be matched against}
\usage{
contains(table, x)
}
\arguments{
\item{x}{the values to be matched}
}
\description{
Test if a vector contains certain values @param table the values to be matched against
}
