% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.evalwith}
\alias{.evalwith}
\title{Convert an object to evaluating environment for list elements Users should not
directly use this function}
\usage{
.evalwith(x)
}
\arguments{
\item{x}{the object}
}
\description{
Convert an object to evaluating environment for list elements Users should not
directly use this function
}
