% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exact-search}
\alias{allEqual}
\alias{allUnequal}
\alias{anyEqual}
\alias{anyUnequal}
\alias{exact-search}
\alias{unidentical}
\title{Exact search functions}
\usage{
allEqual(x, y, ...)

anyEqual(x, y, ...)

allUnequal(x, y, ...)

anyUnequal(x, y, ...)

unidentical(x, y, ...)
}
\arguments{
\item{x}{value}

\item{y}{data}

\item{...}{additional parameters}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Exact search functions
}

