% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fuzzy-search}
\alias{allLike}
\alias{allUnlike}
\alias{anyLike}
\alias{anyUnlike}
\alias{fuzzy-search}
\title{Fuzzy searching functions}
\usage{
allLike(dist, ...)

allUnlike(dist, ...)

anyLike(dist, ...)

anyUnlike(dist, ...)
}
\arguments{
\item{dist}{string distance}

\item{...}{Additional parameter passed to \code{stringdist} functions}
}
\value{
a closure \code{function(src,value)}
}
\description{
Fuzzy searching functions
}

