% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getnames}
\alias{getnames}
\title{Get the names of an object}
\usage{
getnames(x, def = NULL)
}
\arguments{
\item{x}{the object to extract names}

\item{def}{the value to return if the object has \code{NULL} names.
For vectorization purpose, set this to \code{character(1L)}.}
}
\description{
Get the names of an object
}
\details{
This function is used in vectorization when the names of an object
  is to be supplied.  \code{NULL} value will break the vectorization while
  setting \code{def = character(1L)} makes the names vectorizable.
}
