% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.empty}
\alias{is.empty}
\title{Check if an object is empty (has length 0)}
\usage{
is.empty(x)
}
\arguments{
\item{x}{the object}
}
\description{
Check if an object is empty (has length 0)
}
\details{
A \code{NULL} value, zero-length vector or list have length zero,
  which is called empty.
}
