% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.all}
\alias{list.all}
\title{Get whether all list members satisfy the given condition}
\usage{
list.all(.data, cond, na.rm = FALSE)
}
\arguments{
\item{.data}{\code{list}}

\item{cond}{A logical lambda expression}

\item{na.rm}{logical. If true all \code{NA} values are removed}
}
\description{
Get whether all list members satisfy the given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.all(x,type=="B")
list.all(x,mean(unlist(score))>=6)
}
}

