% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.any.R
\name{list.any}
\alias{list.any}
\title{Examine if a condition is true for at least one list element}
\usage{
list.any(.data, cond, na.rm = FALSE)
}
\arguments{
\item{.data}{\code{list}}

\item{cond}{A logical lambda expression}

\item{na.rm}{logical. If true the function returns \code{NA} when \code{NA}
is encountered as \code{cond} is evaluated.}
}
\description{
Examine if a condition is true for at least one list element
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.any(x,type=="B")
list.any(x,mean(unlist(score))>=6)
}
}

