% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.apply}
\alias{list.apply}
\title{Apply a function to each list member (\code{lapply})}
\usage{
list.apply(.data, .fun, ...)
}
\arguments{
\item{.data}{\code{list}}

\item{.fun}{\code{function}}

\item{...}{Additional parameters passed to \code{FUN}.}
}
\description{
Apply a function to each list member (\code{lapply})
}

