% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.cases}
\alias{list.cases}
\title{Get all unique cases by expression for a list}
\usage{
list.cases(.data, expr, simplify = TRUE, ..., sort = TRUE)
}
\arguments{
\item{.data}{\code{list}}

\item{expr}{The expression to evaluate to find cases}

\item{simplify}{\code{logical}. Should the values be simplified
by \code{unlist}?}

\item{...}{Additional parameters passed to \code{unique}}

\item{sort}{\code{logical}. Should the cases be sorted in ascending order?}
}
\description{
Get all unique cases by expression for a list
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.cases(x,type)
list.cases(x,mean(unlist(score)))
}
}

