% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.count}
\alias{list.count}
\title{Count the number of members that meet given condition}
\usage{
list.count(.data, cond)
}
\arguments{
\item{.data}{\code{list}}

\item{cond}{A logical lambda expression.}
}
\description{
Count the number of members that meet given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.count(x,type=="B")
list.count(x,min(unlist(score)) >= 9)
}
}

