% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.do}
\alias{list.do}
\title{Call a function with a list of arguments as provided}
\usage{
list.do(.data, fun, ...)
}
\arguments{
\item{.data}{\code{list}}

\item{fun}{The \code{function} to call}

\item{...}{The additional parameters passed to \code{do.call}}
}
\description{
Call a function with a list of arguments as provided
}
\examples{
\dontrun{
x <- lapply(1:3,function(i) { c(a=i,b=i^2)})
df <- lapply(1:3,function(i) { data.frame(a=i,b=i^2,c=letters[i])})
list.do(x,rbind)
list.do(x,rbind)
as.list(1:10) \%>>\% list.map(x -> list.do(x,rnorm))
}
}

