% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.exclude}
\alias{list.exclude}
\title{Exclude members of a list that meet given condition.}
\usage{
list.exclude(x, cond)
}
\arguments{
\item{x}{\code{list}}

\item{cond}{An \code{expression} that returns a logical value
to exclude items}
}
\description{
Exclude members of a list that meet given condition.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.exclude(x,type=="B")
list.exclude(x,min(score$c1,score$c2) >= 8)
}
}

