% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.expand.R
\name{list.expand}
\alias{list.expand}
\title{Create a list from all combinations of factors}
\usage{
list.expand(...)
}
\arguments{
\item{...}{vectors or lists}
}
\value{
A list of all combinations of the supplied vectors or
lists.
}
\description{
Create a list from all combinations of the supplied
vectors or lists, extending the functionality of
\link{expand.grid} from data frame to list.
}
\examples{
list.expand(x=1:10, y=c("a","b","c"))
list.expand(x=list(c(1,2), c(2,3)), y = c("a","b","c"))
list.expand(
  a=list(list(x=1,y="a"), list(x=2, y="b")),
  b=list(c("x","y"), c("y","z","w")))
}
