% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.extract.R
\name{list.extract}
\alias{list.extract}
\title{Extract a member from a list}
\usage{
list.extract()
}
\description{
Extract a member from a list
}
\examples{
x <- list(a=1, b=2, c=3)
list.extract(x, 1)
list.extract(x, "a")
}

