% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.filter}
\alias{list.filter}
\alias{list.map}
\title{Filter a list by a condition.}
\usage{
list.filter(x, filter = TRUE, item = ".", keep.names = TRUE,
  keep.null = FALSE)

list.map(x, select, item = ".", keep.names = TRUE, keep.null = FALSE)
}
\arguments{
\item{x}{The list to be filtered}

\item{subset}{A logical expression that specifies the filtering condition}

\item{item}{The symbol to represent the list item, \code{.} in default}

\item{keep.names}{Whether to keep the names of list x}

\item{keep.null}{Whether to keep \code{NULL} items in the result}

\item{x}{The list to perform mapping}

\item{select}{An expression that is evaluated for each item
that satisfies the subsetting condition}

\item{item}{The symbol to represent the list item, \code{.} in default}

\item{keep.names}{Whether to keep the names of list x}

\item{keep.null}{Whether to keep \code{NULL} items in the result}
}
\description{
Filter a list by a condition.

Map each member of a list by an expression.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.filter(x,type=="B")
list.filter(x,min(score$c1,score$c2) >= 8)
}
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.filter(x,type=="B")
list.filter(x,min(score$c1,score$c2) >= 8)
}
}

