% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.find}
\alias{list.find}
\title{Find a specific number of members in a list that meeting given condition}
\usage{
list.find(x, ..., keep.names = TRUE, keep.null = FALSE)
}
\arguments{
\item{x}{The list}

\item{...}{Parameters passed to \code{list.findi}}

\item{keep.names}{Whether to keep the names of list x}

\item{keep.null}{Whether to keep \code{NULL} items in the result}
}
\description{
Find a specific number of members in a list that meeting given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.find(x,type=="B",1)
list.find(x,min(score$c1,score$c2) >= 9)
}
}

