% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.findi}
\alias{list.findi}
\title{Find the indices of a number of members in a list that
meet given condition}
\usage{
list.findi(x, cond = TRUE, n = 1, item = ".")
}
\arguments{
\item{x}{The list}

\item{cond}{The condition}

\item{n}{The maximal number of members to find out}

\item{item}{The symbol to represent the list item, \code{.} in default}
}
\description{
Find the indices of a number of members in a list that
meet given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.findi(x,type=="B")
list.findi(x,min(score$c1,score$c2) >= 8)
list.findi(x,min(score$c1,score$c2) <= 8,2)
}
}

