% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.first.R
\name{list.first}
\alias{list.first}
\title{Find the first element that meets a condition}
\usage{
list.first(.data, cond, na.rm = TRUE)
}
\arguments{
\item{.data}{list}

\item{cond}{a logical lambda expression}

\item{na.rm}{\code{TRUE} to ignore \code{NA} values}
}
\description{
Find the first element that meets a condition
}

