% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.group}
\alias{list.group}
\title{Group a list by the value of an expression evaluated for each member.}
\usage{
list.group(.data, ..., sorted = TRUE)
}
\arguments{
\item{.data}{\code{list}}

\item{...}{keys}

\item{sorted}{\code{TRUE} to sort the group keys. Ignored when the key has
multiple entries.}
}
\description{
Group a list by the value of an expression evaluated for each member.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.group(x,type)
list.group(x,mean(unlist(score)))
}
}

