% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.if}
\alias{list.if}
\title{Return a logical vector that indicates if each member of a list
satisfies a given condition}
\usage{
list.if(x, cond, item = ".", keep.names = TRUE)
}
\arguments{
\item{x}{A list}

\item{cond}{A logical expression that specifies the condition}

\item{item}{The symbol to represent the list item, \code{.} in default}

\item{keep.names}{Whether to keep the names of list x}
}
\description{
Return a logical vector that indicates if each member of a list
satisfies a given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.if(x,type=="B")
list.if(x,min(score$c1,score$c2) >= 8)
}
}

