% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.insert.R
\name{list.insert}
\alias{list.insert}
\title{Insert a series of lists at the given index}
\usage{
list.insert(.data, index, ...)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{index}{The index at which the lists are inserted}

\item{...}{A group of lists}
}
\description{
Insert a series of lists at the given index
}
\examples{
\dontrun{
x <- list(p1 = list(type='A',score=list(c1=10,c2=8)),
       p2 = list(type='B',score=list(c1=9,c2=9)),
       p3 = list(type='B',score=list(c1=9,c2=7)))
list.insert(x, 2, p2.1 = list(type='B',score=list(c1=8,c2=9)))
}
}
\seealso{
\code{\link{list.append}}, \code{\link{list.prepend}}
}
