% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.is}
\alias{list.if}
\alias{list.is}
\title{Return a logical vector that indicates if each member of a list
satisfies a given condition}
\usage{
list.is(.data, cond, use.names = TRUE, envir = parent.frame())

list.if(.data, cond, use.names = TRUE, envir = parent.frame())
}
\arguments{
\item{.data}{\code{list}}

\item{cond}{A logical lambda expression}

\item{use.names}{\code{logical} Should the names of \code{.data} be kept?}

\item{envir}{The environment to evaluate mapping function}
}
\description{
Return a logical vector that indicates if each member of a list
satisfies a given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.is(x,type=="B")
list.is(x,min(score$c1,score$c2) >= 8)
}
}

