% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.load}
\alias{list.load}
\title{Load a list from file}
\usage{
list.load(file, type = tools::file_ext(file), ...)
}
\arguments{
\item{file}{The file as input}

\item{type}{The type of input which, by default, is determined
by file extension. Currently supports RData, RDS, JSON, YAML.}

\item{...}{Additional parameters passed to the loader function}
}
\description{
Load a list from file
}
\examples{
\dontrun{
list.load("list.rds")
list.load("list.rdata")
list.load("list.yaml")
list.load("list.json")
}
}

