% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.maps}
\alias{list.maps}
\title{Map multiple lists with an expression}
\usage{
list.maps(expr, ..., envir = parent.frame())
}
\arguments{
\item{expr}{An implicit lambda expression where only \code{.i} and
\code{.name} are defined.}

\item{...}{Named arguments of lists with equal length. The names of the
lists are available as symbols that represent the element for each list.}

\item{envir}{The environment to evaluate mapping function}
}
\description{
Map multiple lists with an expression
}
\examples{
\dontrun{
l1 <- list(p1=list(x=1,y=2), p2=list(x=3,y=4), p3=list(x=1,y=3))
l2 <- list(2,3,5)
list.maps(a$x*b+a$y,a=l1,b=l2)
}
}

