% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.mapv}
\alias{list.mapv}
\title{Map each member of a list by an expression to a vector.}
\usage{
list.mapv(.data, expr, use.names = TRUE)
}
\arguments{
\item{.data}{\code{list}}

\item{expr}{The expression}

\item{use.names}{Should the names of the results be preserved?}
}
\description{
Map each member of a list by an expression to a vector.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.mapv(x,type)
list.mapv(x,min(score$c1,score$c2))
}
}

