% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.names}
\alias{list.names}
\title{Get or set the names of a list by expression}
\usage{
list.names(.data, expr)
}
\arguments{
\item{.data}{the list}

\item{expr}{the expression whose value will be set as the name
for each list element. If missing then the names of the list will be
returned}
}
\description{
Get or set the names of a list by expression
}

