% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.order.R
\name{list.order}
\alias{list.order}
\title{Return the order of each member in a list by expression}
\usage{
list.order(.data, ..., keep.names = FALSE, na.last = TRUE)
}
\arguments{
\item{.data}{\code{list}}

\item{...}{A group of lambda expressions}

\item{keep.names}{Whether to keep the names of \code{x} in the result}

\item{na.last}{The way to deal with \code{NA}s.}
}
\description{
Return the order of each member in a list by expression
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.order(x,type,desc(score$c2))
list.order(x,min(score$c1,score$c2))
list.order(x,min(score$c1,score$c2),keep.names=TRUE)
}
}

