% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.query.R
\name{list.query}
\alias{list.query}
\title{Query a list with JsonPath}
\usage{
list.query(x, query, with = list(), validate = TRUE, ...)
}
\arguments{
\item{x}{a list, JSON string, file path or a web link to a json document.}

\item{query}{a JsonPath string.}

\item{with}{a list of elements to be used in \code{query}.}

\item{validate}{\code{TRUE} to validate the JSON document if \code{x}
is supplied one. Ignored if \code{x} is a list.}

\item{...}{Additional parameters. If \code{x} is supplied a file path
or a web address, then these arguments will be passed to \code{readLines}
that reads the file content.}
}
\description{
Query a list with JsonPath
}
\details{
JsonPath for JSON document is similar with XPath for XML documents.
The function creates an internal V8 Javascript engine, converts
the list object to Javascript DOM object, and calls
\url{https://github.com/s3u/JSONPath} methods to query the DOM object,
and finally converts the result back to list.

The function might be less performant with large list or frequent
calling.
}
\seealso{
\url{https://github.com/s3u/JSONPath}
}

