% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.reverse.R
\name{list.reverse}
\alias{list.reverse}
\title{Reverse a list}
\usage{
list.reverse(.data)
}
\arguments{
\item{.data}{\code{list}}
}
\description{
Reverse a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.reverse(x)
}
}

