% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.serialize}
\alias{list.serialize}
\title{Serialize a list}
\usage{
list.serialize(x, file, type = tolower(tools::file_ext(file)), ...)
}
\arguments{
\item{x}{\code{list}}

\item{file}{The file for output}

\item{type}{The type of serialization, including native serializer and
json serializer, which is by default determined by file extension}

\item{...}{Additional parameters passed to the serializer function}
}
\description{
Serialize a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.serialize(x,"test.dat")
list.serialize(x,"test.json")
}
}

