% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.skip}
\alias{list.skip}
\title{Skip a number of members in a list}
\usage{
list.skip(x, n)
}
\arguments{
\item{x}{The list from which some members are taken}

\item{n}{The number of members to skip}
}
\description{
Skip a number of members in a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.skip(x,1)
list.skip(x,2)
}
}

