% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.subset}
\alias{list.subset}
\title{Subset a list}
\usage{
list.subset(x, value, pattern = FALSE, dist = NA_integer_, ...)
}
\arguments{
\item{x}{list}

\item{value}{names or indices}

\item{pattern}{\code{TRUE} if \code{value} indicates a regular
expression pattern.}

\item{dist}{\code{integer} that indicates the maximal string distance to
\code{value}.}

\item{...}{additional parameters.}
}
\description{
Subset a list
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.subset(x, c("p1","p2"))
list.subset(x, "^p", pattern = TRUE)
list.subset(x, "x1", dist = 1)
}
}

