% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.table}
\alias{list.table}
\title{Generate a table for a list by expression}
\usage{
list.table(.data, ..., table.args = NULL)
}
\arguments{
\item{.data}{\code{list}}

\item{...}{A group of lambda expressions}

\item{table.args}{\code{list}. The additional parameters
passed to \code{table}}
}
\description{
Generate a table for a list by expression
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.table(x,type)
list.table(x,type,c1=score$c1)
list.table(x,type,score$c1,table.args=list(dnn=c("type","c1")))
}
}

