% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.zip}
\alias{list.zip}
\title{Combine multiple lists element-wisely.}
\usage{
list.zip(..., use.argnames = TRUE, use.names = TRUE)
}
\arguments{
\item{...}{\code{list}s}

\item{use.argnames}{\code{logical}. Should the names of the
arguments be used as the names of list items?}

\item{use.names}{\code{logical}. Should the names of the first
argument be used as the zipped list?}
}
\description{
Combine multiple lists element-wisely.
}
\examples{
\dontrun{
x <- list(1,2,3)
y <- list("x","y","z")
list.zip(num=x,sym=y)
}
}

