% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyweather}
\alias{nyweather}
\title{New York hourly weather data}
\format{See \url{http://openweathermap.org/weather-data#current}}
\source{
\url{http://api.openweathermap.org/data/2.5/history/city?q=New York}
}
\usage{
nyweather
}
\description{
A non-tabular data of the hourly weather conditions of the New York City
from 2013-01-01 to 2013-03-01.
}
\details{
Fetch date: 2014-11-23.

Processed by rlist.
}
\keyword{datasets}

