% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlist.R
\docType{package}
\name{rlist-package}
\alias{rlist-package}
\title{The rlist package}
\description{
The rlist package
}
\details{
rlist is a set of tools for working with list objects. Its goal
is to make it easier to work with lists by providing a wide range
of functions that operate on non-tabular data stored in them.

The package provides a set of functions for data manipulation with
list objects, including mapping, filtering, grouping, sorting,
updating, searching, and other useful functions. Most functions
are designed to be pipeline friendly so that data processing with
lists can be chained.

rlist Tutorial (\url{http://renkun.me/rlist-tutorial}) is a complete guide to rlist.
}

