% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_argnames}
\alias{set_argnames}
\title{Make names for unnamed symbol arguments}
\usage{
set_argnames(args, data = args)
}
\arguments{
\item{args}{the unevaluated argument list}

\item{data}{the list to be named (\code{args} by default)}
}
\description{
Make names for unnamed symbol arguments
}
\details{
The elements of an unevaluated list of arguments may or may not
have names as given by user. For example, \code{list.select} requires user
to specify the fields to select. These fields are unevaluated arguments,
some of which are symbols and others are calls. For the symbols, it is natural
to make the resulted lists to have the same name for the particular arguments.
}
