% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.list}
\alias{summary.list}
\title{Summarize a list.}
\usage{
\method{summary}{list}(object, verbose = show.info | show.data,
  show.info = TRUE, show.data = 4, ...)
}
\arguments{
\item{object}{The list to summarize}

\item{verbose}{Output list information and data briefing}

\item{show.info}{Output list information}

\item{show.data}{Output list data briefing as yaml}

\item{...}{Additional parameters passed to \code{yaml::as.yaml}}
}
\value{
List information and data briefing in \code{yaml} format will be printed and \code{summary.default} result will be returned
}
\description{
Summarize a list.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
summary(x)
summary(x,show.info=FALSE)
summary(x,show.data=TRUE)
summary(x,show.data=2)
summary(x,indent=4)
}
}

