% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transform.list}
\alias{transform.list}
\title{Transform a list by modifying its elements.}
\usage{
\method{transform}{list}(`_data`, ..., item = ".", keep.names = TRUE,
  keep.null = FALSE, keep.val.null = FALSE)
}
\arguments{
\item{_data}{The list to be transformed}

\item{...}{The elements to transform}

\item{item}{The symbol to represent the list item, \code{.} in default}

\item{keep.names}{Whether to keep the names of list x}

\item{keep.null}{Whether to keep \code{NULL} items in the result}

\item{keep.val.null}{Whether to keep \code{NULL} values in the transformed lsit item}
}
\description{
Transform a list by modifying its elements.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
transform(x,high=max(score$c1,score$c2),low=min(score$c1,score$c2))
transform(x,exams=length(score))
transform(x,grade=ifelse(type=="A",score$c1,score$c2))
transform(x,score=list(min=0,max=10))
}
}

