% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exact-search}
\alias{equal}
\alias{exact-search}
\alias{exclude}
\alias{include}
\alias{unequal}
\alias{unidentical}
\title{Exact search functions}
\usage{
equal(x, y)

unequal(x, y)

unidentical(x, y, ...)

include(x, y)

exclude(x, y)
}
\arguments{
\item{x}{value}

\item{y}{data}

\item{...}{additional parameters}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Exact search functions
}

