% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fuzzy-search}
\alias{fuzzy-search}
\alias{like}
\alias{unlike}
\title{Fuzzy search functions}
\usage{
like(dist, ...)

unlike(dist, ...)
}
\arguments{
\item{dist}{string distance}

\item{...}{additional parameter passed to
\code{stringdist} functions}
}
\value{
a closure \code{function(src,value)}
}
\description{
Fuzzy search functions
}

