% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.append.R
\name{list.append}
\alias{list.append}
\title{Append a list}
\usage{
list.append(.data, ...)
}
\arguments{
\item{.data}{The original list}

\item{...}{A \code{vector} or \code{list} to append after \code{x}}
}
\description{
Append a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.append(x,d=4,e=5)
list.append(x,d=4,f=c(2,3))
}
}

