% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.append}
\alias{list.append}
\title{Append a list}
\usage{
list.append(x, ...)
}
\arguments{
\item{x}{The original list}

\item{...}{A vector or list that is to be appended after \code{x}}
}
\description{
Append a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.append(x,c(d=4,e=5))
list.append(x,list(d=4,f=c(2,3)))
}
}

