% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.insert.R
\name{list.append}
\alias{list.append}
\title{Append elements to a list}
\usage{
list.append(.data, ...)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{...}{A \code{vector} or \code{list} to append after \code{x}}
}
\description{
Append elements to a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.append(x,d=4,e=5)
list.append(x,d=4,f=c(2,3))
}
}
\seealso{
\code{\link{list.prepend}}, \code{\link{list.insert}}
}
