% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.do.R
\name{list.apply}
\alias{list.apply}
\title{Apply a function to each list element (\code{lapply})}
\usage{
list.apply(.data, .fun, ...)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{.fun}{\code{function}}

\item{...}{Additional parameters passed to \code{FUN}.}
}
\description{
Apply a function to each list element (\code{lapply})
}
