% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.do.R
\name{list.cbind}
\alias{list.cbind}
\title{Bind all list elements by column}
\usage{
list.cbind(.data)
}
\arguments{
\item{.data}{\code{list}}
}
\description{
Bind all list elements by column
}
\examples{
x <- list(data.frame(i=1:5,x=rnorm(5)),
   data.frame(y=rnorm(5),z=rnorm(5)))
list.cbind(x)
}
\seealso{
\code{\link{list.cbind}}, \code{\link{list.stack}}
}

