% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.cbind}
\alias{list.cbind}
\title{Bind all list members by column}
\usage{
list.cbind(.data)
}
\arguments{
\item{.data}{\code{list}}
}
\description{
Bind all list members by column
}
\examples{
\dontrun{
x <- list(data.frame(i=1:5,x=rnorm(5)),
   data.frame(y=rnorm(5),z=rnorm(5)))
list.cbind(x)
}
}

