% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.clean}
\alias{list.clean}
\title{Clean a list by a function}
\usage{
list.clean(.data, fun = is.null)
}
\arguments{
\item{.data}{The list}

\item{fun}{The function for cleaning, by default \code{is.null}.}
}
\description{
Clean a list by a function
}
\examples{
\dontrun{
x <- list(a=NULL,b=NULL,c=NULL,d=1,e=2)
list.clean(x)
}
}

