% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.clean}
\alias{list.clean}
\title{Clean a list by a function}
\usage{
list.clean(.data, fun = is.null, recursive = FALSE)
}
\arguments{
\item{.data}{\code{list}}

\item{fun}{A logical \code{function} for clean}

\item{recursive}{\code{logical}. Should the list be
cleaned recursively?}
}
\description{
Clean a list by a function
}
\examples{
\dontrun{
x <- list(a=NULL,b=NULL,c=NULL,d=1,e=2)
list.clean(x)
}
}

