% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.common.R
\name{list.common}
\alias{list.common}
\title{Get all common cases by expression for a list}
\usage{
list.common(.data, expr)
}
\arguments{
\item{.data}{\code{list}}

\item{expr}{An anonymous (or "lambda") expression to determine common cases. If one
is not specified, \code{list.common} simply returns all identical sub-elements within lists.}
}
\description{
Get all common cases by expression for a list
}
\examples{
x <- list(c('a','b','c'),c('a','b'),c('b','c'))
list.common(x, .)
x <- list(p1 = list(type='A',score=list(c1=10,c2=8)),
       p2 = list(type='B',score=list(c1=9,c2=9)),
       p3 = list(type='B',score=list(c1=9,c2=7)))
list.common(x,type)
list.common(x,names(score))

foo <- list(x = LETTERS[1:3], y = LETTERS[3:5])
list.common(foo)
}
