% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.remove.R
\name{list.exclude}
\alias{list.exclude}
\title{Exclude members of a list that meet given condition.}
\usage{
list.exclude(.data, cond)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{cond}{A logical lambda expression to exclude items}
}
\description{
Exclude members of a list that meet given condition.
}
\examples{
x <- list(p1 = list(type='A',score=list(c1=10,c2=8)),
       p2 = list(type='B',score=list(c1=9,c2=9)),
       p3 = list(type='B',score=list(c1=9,c2=7)))
list.exclude(x, type=='B')
list.exclude(x, min(score$c1,score$c2) >= 8)
}
