% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.extract.R
\name{list.extract}
\alias{list.extract}
\title{Extract an element from a list or vector}
\usage{
list.extract()
}
\description{
Extract an element from a list or vector
}
\examples{
x <- list(a=1, b=2, c=3)
list.extract(x, 1)
list.extract(x, 'a')
}
