% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.filter.R
\name{list.filter}
\alias{list.filter}
\title{Filter a list by a condition.}
\usage{
list.filter(.data, ...)
}
\arguments{
\item{.data}{\code{list}}

\item{...}{logical conditions}
}
\description{
Filter a list by a condition.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.filter(x,type=="B")
list.filter(x,min(score$c1,score$c2) >= 8)
list.filter(x, type=="B", score$c2 >= 8)
}
}

