% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.find}
\alias{list.find}
\title{Find a specific number of members in a list that meeting given condition}
\usage{
list.find(.data, cond, n = 1L, envir = parent.frame())
}
\arguments{
\item{.data}{\code{list}}

\item{cond}{A logical lambda expression}

\item{n}{The number of items to find}

\item{envir}{The environment to evaluate mapping function}
}
\description{
Find a specific number of members in a list that meeting given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.find(x,type=="B",1)
list.find(x,min(score$c1,score$c2) >= 9)
}
}

