% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.find.R
\name{list.find}
\alias{list.find}
\title{Find a specific number of elements in a list or vector
satisfying a given condition}
\usage{
list.find(.data, cond, n = 1L)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{cond}{A logical lambda expression}

\item{n}{The number of items to find. (\code{n = 1L} by default)}
}
\value{
A list or vector of at most \code{n} elements in \code{.data}
found to satisfy \code{cond}.
}
\description{
Find a specific number of elements in a list or vector
satisfying a given condition
}
\examples{
x <- list(p1 = list(type='A',score=list(c1=10,c2=8)),
       p2 = list(type='B',score=list(c1=9,c2=9)),
       p3 = list(type='B',score=list(c1=9,c2=7)))
list.find(x, type=='B', 1)
list.find(x, min(score$c1,score$c2) >= 9)
}
