% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.find.R
\name{list.findi}
\alias{list.findi}
\title{Find the indices of a number of members in a list that
meet given condition}
\usage{
list.findi(.data, cond, n = 1L)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{cond}{A logical lambda expression}

\item{n}{The maximal number of members to find out}
}
\description{
Find the indices of a number of members in a list that
meet given condition
}
\examples{
x <- list(p1 = list(type='A',score=list(c1=10,c2=8)),
       p2 = list(type='B',score=list(c1=9,c2=9)),
       p3 = list(type='B',score=list(c1=9,c2=7)))
list.findi(x, type=='B')
list.findi(x, min(score$c1,score$c2) >= 8)
list.findi(x, min(score$c1,score$c2) <= 8, n = 2)
}

