% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.flatten.R
\name{list.flatten}
\alias{list.flatten}
\title{Flatten a list to only one level}
\usage{
list.flatten(x, use.names = TRUE)
}
\arguments{
\item{x}{\code{list}}

\item{use.names}{\code{logical}. Should the names of \code{x} be kept?}
}
\description{
Flatten a list to only one level
}
\examples{
\dontrun{
p <- list(a=1,b=list(b1=2,b2=3),c=list(c1=list(c11="a",c12="x"),c2=3))
list.flatten(p)
}
}

