% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.iter.R
\name{list.iter}
\alias{list.iter}
\title{Iterate a list by evaluating an expression on
each list member.}
\usage{
list.iter(.data, expr)
}
\arguments{
\item{.data}{\code{list}}

\item{expr}{A lambda expression}
}
\value{
\code{invisible(.data)}
}
\description{
Iterate a list by evaluating an expression on
each list member.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.iter(x,cat(paste(type,"\\n")))
list.iter(x,cat(str(.)))
}
}

