% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.load}
\alias{list.load}
\title{Load a list from file}
\usage{
list.load(file, type = tolower(tools::file_ext(file)), ...)
}
\arguments{
\item{file}{The file as input}

\item{type}{The type of input which is by default determined
by file extension}

\item{...}{Additional parameters passed to the loader function}
}
\description{
Load a list from file
}
\examples{
\dontrun{
list.load("list.rdata")
list.load("list.yaml")
list.load("list.json")
}
}

