% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.map.R
\name{list.map}
\alias{list.map}
\title{Map each element in a list or vector by an expression.}
\usage{
list.map(.data, expr)
}
\arguments{
\item{.data}{a \code{list} or \code{vector}}

\item{expr}{A lambda expression}
}
\value{
A \code{list} in which each element is mapped by \code{expr} in \code{.data}
}
\description{
Map each element in a list or vector by an expression.
}
\examples{
x <- list(p1 = list(type='A',score=list(c1=10,c2=8)),
       p2 = list(type='B',score=list(c1=9,c2=9)),
       p3 = list(type='B',score=list(c1=9,c2=7)))
list.map(x, type)
list.map(x, min(score$c1,score$c2))
}
\seealso{
\code{\link{list.mapv}}
}
