% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.map.R
\name{list.maps}
\alias{list.maps}
\title{Map multiple lists with an expression}
\usage{
list.maps(expr, ...)
}
\arguments{
\item{expr}{An implicit lambda expression where only \code{.i} and
\code{.name} are defined.}

\item{...}{Named arguments of lists with equal length. The names of the
lists are available as symbols that represent the element for each list.}
}
\description{
Map multiple lists with an expression
}
\examples{
\dontrun{
l1 <- list(p1=list(x=1,y=2), p2=list(x=3,y=4), p3=list(x=1,y=3))
l2 <- list(2,3,5)
list.maps(a$x*b+a$y,a=l1,b=l2)
list.maps(..1$x*..2+..1$y,l1,l2)
}
}
