% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.merge.R
\name{list.merge}
\alias{list.merge}
\title{Merge a series of named lists}
\usage{
list.merge(...)
}
\arguments{
\item{...}{named lists}
}
\description{
Merge a series of named lists
}
\examples{
\dontrun{
l1 <- list(a=1,b=list(x=1,y=1))
l2 <- list(a=2,b=list(z=2))
l3 <- list(a=2,b=list(x=3))
list.merge(l1,l2,l3)
}
}

