% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.merge}
\alias{list.merge}
\title{Merge two lists}
\usage{
list.merge(x, val, keep.null = FALSE)
}
\arguments{
\item{x}{The original list}

\item{val}{A vector or list that is to be appended after \code{x}}

\item{keep.null}{Whether to keep \code{NULL} items}
}
\description{
Merge two lists
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=list(x=1,y=2))
list.merge(x,list(b=5))
list.merge(x,list(c=list(z=3)))
}
}

