% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.merge}
\alias{list.merge}
\title{Merge a series of lists}
\usage{
list.merge(...)
}
\arguments{
\item{...}{A group of lists}
}
\description{
Merge a series of lists
}
\examples{
\dontrun{
l1 <- list(a=1,b=list(x=1,y=1))
l2 <- list(a=2,b=list(z=2))
l3 <- list(a=2,b=list(x=3))
list.merge(l1,l2,l3)
}
}

