% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.names}
\alias{list.names}
\title{Get or set the names of a list by expression}
\usage{
list.names(.data, expr)
}
\arguments{
\item{.data}{the list}

\item{expr}{the expression whose value will be set as the name
for each list element. If missing then the names of the list will be
returned}
}
\description{
Get or set the names of a list by expression
}
\examples{
list.names(c(1,2,3))
list.names(c(a=1,b=2,c=3))
list.names(c(1,2,3),letters[.])
list.names(list(list(name="A",value=10),list(name="B",value=20)), name)
}

