% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.names.R
\name{list.names}
\alias{list.names}
\title{Get or set the names of a list by expression}
\usage{
list.names(.data, expr)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{expr}{the expression whose value will be set as the name
for each list element. If missing then the names of the list will be
returned. If \code{NULL} then the names of the list will be removed.}
}
\description{
Get or set the names of a list by expression
}
\examples{
list.names(c(1,2,3))
list.names(c(a=1,b=2,c=3))
list.names(c(1,2,3),letters[.])
list.names(list(list(name='A',value=10),list(name='B',value=20)), name)
}
