% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.prepend}
\alias{list.prepend}
\title{Prepend a list}
\usage{
list.prepend(.data, ...)
}
\arguments{
\item{.data}{\code{list}}

\item{...}{A vector or list that is to be prepended before \code{x}}
}
\description{
Prepend a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.prepend(x,d=4,e=5)
list.prepend(x,d=4,f=c(2,3))
}
}

