% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.insert.R
\name{list.prepend}
\alias{list.prepend}
\title{Prepend elements to a list}
\usage{
list.prepend(.data, ...)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{...}{The \code{vector} or \code{list} to prepend before \code{x}}
}
\description{
Prepend elements to a list
}
\examples{
x <- list(a=1,b=2,c=3)
list.prepend(x, d=4, e=5)
list.prepend(x, d=4, f=c(2,3))
}
\seealso{
\code{\link{list.append}}, \code{\link{list.insert}}
}
