% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.rbind}
\alias{list.rbind}
\title{Bind all list members by row}
\usage{
list.rbind(.data)
}
\arguments{
\item{.data}{\code{list}}
}
\description{
Bind all list members by row
}
\examples{
\dontrun{
x <- lapply(1:3,function(i) { c(a=i,b=i^2)})
df <- lapply(1:3,function(i) { data.frame(a=i,b=i^2,c=letters[i])})
list.rbind(x)
list.rbind(df)
}
}

