% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.reverse.R
\name{list.reverse}
\alias{list.reverse}
\title{Reverse a list}
\usage{
list.reverse(.data)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}
}
\description{
Reverse a list
}
\examples{
x <- list(a=1,b=2,c=3)
list.reverse(x)
}
