% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.sample}
\alias{list.sample}
\title{Sample a list}
\usage{
list.sample(.data, size, replace = FALSE, weight = 1, prob = NULL)
}
\arguments{
\item{.data}{list{code}}

\item{size}{The size the sample to take}

\item{replace}{Should sampling be with replacement?}

\item{weight}{An expression to determine the weight of each list member,
which does only takes effectt if \code{prob} is \code{NULL}.}

\item{prob}{A vector of probability weights for obtaining the elements of the
list being sampled.}
}
\description{
Sample a list
}
\examples{
\dontrun{
x <- lapply(1:3,function(i) { c(a=i,b=i^2)})
df <- lapply(1:3,function(i) { data.frame(a=i,b=i^2,c=letters[i])})
list.rbind(x)
list.rbind(df)
}
}

