% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.search}
\alias{list.search}
\title{Search a list recusively by a value}
\usage{
list.search(.data, fun, value, ..., classes = class(value), unlist = FALSE)
}
\arguments{
\item{.data}{\code{list}}

\item{fun}{A logical comparer \code{function}.

    Exact search comparers

    \code{identical}, \code{allEqual}, \code{anyEqual},
    \code{unidentical}, \code{allUnequal}, \code{anyUnequal}

    Fuzzy search comparers

    \code{anyLike(dist)}, \code{allLike(dist)},
    \code{anyUnlike(dist)}, \code{allUnlike(dist)} where
    \code{dist} is the maximum or minimum string distance.}

\item{value}{The value to search}

\item{...}{Additional parameters passed to \code{fun}}

\item{classes}{A character vector of class names that restrict the search. By default, the search range is restrcited to the class of \code{value}, so that types are strictly distinguished, that is, searching numeric \code{1} does not include comparing integer \code{1L}. To broader the search range, add more values to \code{classes} such as \code{classes = c("numeric","integer")}.}

\item{unlist}{\code{logical} Should the result be unlisted?}
}
\description{
Search a list recusively by a value
}
\examples{
\dontrun{
# Exact search

x <- list(p1 = list(type="A",score=c(c1=9)),
       p2 = list(type=c("A","B"),score=c(c1=8,c2=9)),
       p3 = list(type=c("B","C"),score=c(c1=9,c2=7)),
       p4 = list(type=c("B","C"),score=c(c1=8,c2=NA)))

## Search identical values
list.search(x, identical, "A")
list.search(x, identical, c("A","B"))
list.search(x, identical, c(10,8))
list.search(x, identical, c(c1=10,c2=8))

## Search all equal values
list.search(x, allEqual, 9)
list.search(x, allEqual, c(10,8))

# Fuzzy search

data <- list(
  p1 = list(name="Ken",age=24),
  p2 = list(name="Kent",age=26),
  p3 = list(name="Sam",age=24),
  p4 = list(name="Keynes",age=30),
  p5 = list(name="Kwen",age=31)
)

list.search(data,anyLike(1), "Ken")
list.search(data,allLike(1), "Ken")

data <- list(
  p1 = list(name=c("Ken", "Ren"),age=24),
  p2 = list(name=c("Kent", "Potter"),age=26),
  p3 = list(name=c("Sam", "Lee"),age=24),
  p4 = list(name=c("Keynes", "Bond"),age=30),
  p5 = list(name=c("Kwen", "Hu"),age=31))

list.search(data, allLike(1), "Ken")
list.search(data, anyLike(1), "Ken")
list.search(data, allUnlike(1), "Ken")
list.search(data, anyUnlike(1), "Ken")
}
}

