% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.search}
\alias{list.search}
\title{Search a list recusively by a value}
\usage{
list.search(.data, value, fun = equal, ..., classes = class(value),
  unlist = FALSE)
}
\arguments{
\item{.data}{\code{list}}

\item{value}{The value to search}

\item{fun}{A logical \code{function} to compare value and list values}

\item{...}{Additional parameters passed to \code{fun}}

\item{classes}{A character vector of class names that restrict the search. By default, the search range is restrcited to the class of \code{value}, so that types are strictly distinguished, that is, searching numeric \code{1} does not include comparing integer \code{1L}. To broader the search range, add more values to \code{classes} such as \code{classes = c("numeric","integer")}.}

\item{unlist}{\code{logical} Should the result be unlisted?}
}
\description{
Search a list recusively by a value
}
\examples{
\dontrun{
# Exact search
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.search(x, 9)

# Case search
data <- list(
  p1 = list(x=c("A","B"),y=c("A","C")),
  p2 = list(q=c("B","C"),p=c("A","C","B")),
  p3 = list(m=list(c("A","B"),c("B","C")),n=c("A"))
)

list.search(data,"A","\%in\%")

# Fuzzy search with \\code{stringdist} package
data <- list(
  p1 = list(name="Ken",age=24),
  p2 = list(name="Kent",age=26),
  p3 = list(name="Sam",age=24),
  p4 = list(name="Keynes",age=30),
  p5 = list(name="Kwen",age=31)
)

list.search(data,"Ken",anyLike(1))
list.search(data,"Ken",allLike(1))

x <- list(
  p1 = list(name=c("Ken", "Ren"),age=24),
  p2 = list(name=c("Kent", "Potter"),age=26),
  p3 = list(name=c("Sam", "Lee"),age=24),
  p4 = list(name=c("Keynes", "Bond"),age=30),
  p5 = list(name=c("Kwen", "Hu"),age=31))
list.search(data,"Ken",allUnlike(1))
list.search(data,"Ken",anyUnlike(2))
}
}

