% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.skip}
\alias{list.skip}
\title{Skip a number of members in a list}
\usage{
list.skip(.data, n)
}
\arguments{
\item{.data}{\code{list}}

\item{n}{\code{integer}. The number to skip}
}
\description{
Skip a number of members in a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.skip(x,1)
list.skip(x,2)
}
}

