% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.skip.R
\name{list.skip}
\alias{list.skip}
\title{Skip a number of members in a list}
\usage{
list.skip(.data, n)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{n}{\code{integer}. The number of elements to skip}
}
\description{
Skip a number of members in a list
}
\examples{
x <- list(a=1,b=2,c=3)
list.skip(x, 1)
list.skip(x, 2)
}

