% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.skip.R
\name{list.skip}
\alias{list.skip}
\title{Skip a number of elements}
\usage{
list.skip(.data, n)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{n}{\code{integer}. The number of elements to skip}
}
\description{
Skip the first \code{n} elements of a list or vector and
return the remaining elements if any.
}
\examples{
x <- list(a=1,b=2,c=3)
list.skip(x, 1)
list.skip(x, 2)
}
\seealso{
\code{\link{list.skipWhile}}, \code{\link{list.take}},
\code{\link{list.takeWhile}}
}
