% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.skipWhile.R
\name{list.skipWhile}
\alias{list.skipWhile}
\title{Skip members until the given condition is broken}
\usage{
list.skipWhile(.data, cond)
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{cond}{A logical lambda expression}
}
\description{
Skip members until the given condition is broken
}
\examples{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.skipWhile(x, type=="A")
list.skipWhile(x, min(score$c1,score$c2) >= 8)
}

