% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.stack}
\alias{list.stack}
\title{Stack all list members to construct a \code{data.frame}}
\usage{
list.stack(.data)
}
\arguments{
\item{.data}{\code{list} of \code{vector}s, \code{list}s,
or \code{data.frame}s.}
}
\description{
Stack all list members to construct a \code{data.frame}
}
\examples{
\dontrun{
x <- lapply(1:3, function(i) { list(a=i,b=i^2) })
list.stack(x)

x <- lapply(1:3, function(i) { list(a=i,b=i^2,c=letters[i])})
list.stack(x)

x <- lapply(1:3, function(i) { data.frame(a=i,b=i^2,c=letters[i]) })
list.stack(x)
}
}

