% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.table.R
\name{list.table}
\alias{list.table}
\title{Generate a table for a list by expression}
\usage{
list.table(.data, ..., table.args = list(useNA = "ifany"))
}
\arguments{
\item{.data}{A \code{list} or \code{vector}}

\item{...}{A group of lambda expressions. If missing,
\code{table} will be directly called upon \code{.data} with
\code{table.args}.}

\item{table.args}{\code{list}. The additional parameters
passed to \code{table}}
}
\description{
Generate a table for a list by expression
}
\examples{
x <- list(p1 = list(type='A',score=list(c1=10,c2=8)),
       p2 = list(type='B',score=list(c1=9,c2=9)),
       p3 = list(type='B',score=list(c1=9,c2=7)))
list.table(x, type)
list.table(x, type, c1 = score$c1)
list.table(x, type, score$c1, table.args = list(dnn=c('type','c1')))
}
