% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.take}
\alias{list.take}
\title{Take a number of members from a list}
\usage{
list.take(.data, n, force = FALSE)
}
\arguments{
\item{.data}{\code{list}}

\item{n}{The number of members to take out}

\item{force}{\code{TRUE} disables the length check}
}
\description{
Take a number of members from a list
}
\examples{
\dontrun{
x <- list(a=1,b=2,c=3)
list.take(x,1)
list.take(x,10)
}
}

