% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.take.R
\name{list.take}
\alias{list.take}
\title{Take a number of elements}
\usage{
list.take(.data, n, force = FALSE)
}
\arguments{
\item{.data}{\code{list} or \code{vector}}

\item{n}{\code{integer}. The number of elements to take}

\item{force}{\code{TRUE} to disable the length check}
}
\description{
Take the first \code{n} elements out from a list or
vector.
}
\examples{
x <- list(a=1,b=2,c=3)
list.take(x,1)
list.take(x,10)
}
\seealso{
\code{\link{list.takeWhile}}, \code{\link{list.skip}},
\code{\link{list.skipWhile}}
}
