% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.takeWhile.R
\name{list.takeWhile}
\alias{list.takeWhile}
\title{Take out members until the given condition is broken}
\usage{
list.takeWhile(.data, cond)
}
\arguments{
\item{.data}{\code{list}}

\item{cond}{A logical lambda expression}
}
\description{
Take out members until the given condition is broken
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.takeWhile(x,type=="B")
list.takeWhile(x,min(score$c1,score$c2) >= 8)
}
}

