% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.takeWhile}
\alias{list.takeWhile}
\title{Take out members until the given condition is broken}
\usage{
list.takeWhile(x, cond = TRUE, keep.names = TRUE)
}
\arguments{
\item{x}{The list}

\item{cond}{The condition}

\item{keep.names}{Whether to keep the names of list x}
}
\description{
Take out members until the given condition is broken
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.takeWhile(x,type=="B")
list.takeWhile(x,min(score$c1,score$c2) >= 8)
}
}

