% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.ungroup}
\alias{list.ungroup}
\title{Ungroup a list}
\usage{
list.ungroup(.data, sort.names = FALSE)
}
\arguments{
\item{.data}{\code{list}}

\item{sort.names}{\code{logical}. Should the members be sorted
after ungrouping?}
}
\description{
Ungroup a list
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
xg <- list.group(x,type)
list.ungroup(xg)
}
}

