% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.which}
\alias{list.which}
\title{Return a integer vector of the indices of list members that satisfy
given condition}
\usage{
list.which(.data, cond, envir = parent.frame())
}
\arguments{
\item{.data}{\code{list}}

\item{cond}{A logical lambda expression}

\item{envir}{The environment to evaluate mapping function}
}
\description{
Return a integer vector of the indices of list members that satisfy
given condition
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=list(c1=10,c2=8)),
       p2 = list(type="B",score=list(c1=9,c2=9)),
       p3 = list(type="B",score=list(c1=9,c2=7)))
list.which(x,type=="B")
list.which(x,min(score$c1,score$c2) >= 8)
}
}

