% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyweather}
\alias{nyweather}
\title{New York hourly weather data}
\format{
See \url{https://openweathermap.org/weather-data}
}
\usage{
nyweather
}
\description{
A non-tabular data of the hourly weather conditions of the New York City
from 2013-01-01 to 2013-03-01.
}
\details{
Fetch date: 2014-11-23.

Processed by rlist.

To retrieve the data, please visit \url{https://openweathermap.org/api} for
API usage.
}
\keyword{datasets}
