% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rlist.R
\docType{data}
\name{Rlist}
\alias{Rlist}
\title{Rlist}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x00000000091942a8> 
 - attr(*, "name")= chr "Rlist_generator"
}}
\usage{
Rlist
}
\description{
Rlist
}
\keyword{datasets}

