% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.list}
\alias{subset.list}
\title{Return subsets of a list which meet conditions.}
\usage{
\method{subset}{list}(list, subset = TRUE, select = NULL, ...)
}
\arguments{
\item{list}{The list to be subsetted}

\item{subset}{A logical expression that specifies the subsetting condition}

\item{select}{An expression that is evaluated for each item that satisfies the subsetting condition}

\item{...}{Additional arguments}
}
\description{
Return subsets of a list which meet conditions.
}
\examples{
\dontrun{
x <- list(p1 = list(type="A",score=10),
       p2 = list(type="B",score=8),
       p3 = list(type="B",score=7))
subset(x,type=="B")
subset(x,select=score)
subset(x,type=="B" & score>=8,paste(type,score,sep=":"))
}
}

